+++ Magick.NET.Native/OpenCL/OpenCLDevice.h
@@ -1,4 +1,4 @@
-﻿//=================================================================================================
+//=================================================================================================
 // Copyright 2013-2016 Dirk Lemstra <https://magick.codeplex.com/>
 //
 // Licensed under the ImageMagick License (the "License"); you may not use this file except in 
@@ -11,29 +11,13 @@
 // express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
 //=================================================================================================
+#pragma once
 
-using System;
-using System.IO;
+MAGICK_NET_EXPORT MagickBooleanType OpenCLDevice_IsEnabled_Get(const MagickCLDevice);
+MAGICK_NET_EXPORT void OpenCLDevice_IsEnabled_Set(const MagickCLDevice, const MagickBooleanType);
 
-namespace ImageMagick
-{
-  internal static partial class FileHelper
-  {
-    public static string CheckForBaseDirectory(string fileName)
-    {
-      if (string.IsNullOrEmpty(fileName))
-        return fileName;
+MAGICK_NET_EXPORT const char *OpenCLDevice_Name_Get(const MagickCLDevice);
 
-      if (fileName.Length < 2 || fileName[0] != '~')
-        return fileName;
+MAGICK_NET_EXPORT const MagickCLDeviceType OpenCLDevice_Type_Get(const MagickCLDevice);
 
-      return AppDomain.CurrentDomain.BaseDirectory + fileName.Substring(1);
-    }
-
-    public static void Delete(FileInfo file)
-    {
-      if (file.Exists)
-        file.Delete();
-    }
-  }
-}
+MAGICK_NET_EXPORT const char *OpenCLDevice_Version_Get(const MagickCLDevice);